+++ osx/build.sh
@@ -3,7 +3,7 @@
 # The above is very useful during debugging.
 # 
 # **************************************************
-# Build script for KA-Lite using Packages and PyRun.
+# Build script for KA-Lite using Packages and virtualenvwrapper.
 #
 # Environment Variable/s:
 # . IS_KALITE_RELEASE == must be set to sign the .app and .pkg packages
@@ -19,16 +19,20 @@
 # 2. Check for valid arguments in terminal.
 # 3. Create temporary directory `temp`.
 # 4. Download the contentpacks/en.zip.
-# 5. Get Github source, optionally use argument for the Github .zip URL, extract, and rename it to `ka-lite`.
-# 6. Get Pyrun, then insert path to the Pyrun binaries in $PATH so Pyrun's python runs first instead of the system python.
-# 7. Upgrade Pyrun's Pip
-# 8. Run `pip install -r requirements_dev.txt` to install the Makefile executables.
-# 9. Run `make dist` for assets and docs.
-# 10. Run `pyrun setup.py install --static` inside the `temp/ka-lite/` directory.
-# 11. Build the Xcode project.
-# 12. Codesign the built .app if running on build server.
-# 13. Run Packages script to build the .pkg.
-#
+# 5. Download python installer.
+# 6. Get Github source, optionally use argument for the Github .zip URL, extract, and rename it to `ka-lite`.
+# 7. Install and create virtualenv.
+# 8. Upgrade Python Pip
+# 9. Run `pip install -r requirements_dev.txt` to install the Makefile executables.
+# 10. Installing PEX to create kalite PEX file
+# 11. Update KA Lite installer version.
+# 12. Run `make dist` for assets and docs.
+# 13. Build the Xcode project.
+# 14. Code-sign the built .app if running on build server.
+# 15. Run Packages script to build the .pkg.
+# 16. Build the dmg file.
+
+
 # REF: Bash References
 # . http://www.peterbe.com/plog/set-ex
 # . http://stackoverflow.com/questions/3601515/how-to-check-if-a-variable-is-set-in-bash
@@ -37,23 +41,25 @@
 # . http://stackoverflow.com/questions/592620/check-if-a-program-exists-from-a-bash-script
 # . http://stackoverflow.com/questions/2924422/how-do-i-determine-if-a-web-page-exists-with-shell-scripting/20988182#20988182
 # . http://stackoverflow.com/questions/2751227/how-to-download-source-in-zip-format-from-github/18222354#18222354
-#
+# . http://stackoverflow.com/questions/592620/check-if-a-program-exists-from-a-bash-script
+# 
 # 
 # MUST: test the signed .pkg or .app on a target Mac with:
 # . pkgutil --check-signature KA-Lite.pkg -- or;
 # . pkgutil --check-signature KA-Lite.app -- or;
 # . spctl --assess --type install KA-Lite.pkg
 
-
-echo "KA-Lite OS X build script for version 0.16.x and above."
-
 STEP=0
-STEPS=13
+STEPS=16
 
 # TODO(cpauya): get version from `ka-lite/kalite/version.py`
-VERSION="0.16"
+# Set the default value to `develop` as suggested by [@benjaoming](https://github.com/learningequality/ka-lite-installers/pull/433#discussion_r96399812), so we can use the VERSION environment in bamboo settings. 
+VERSION=${VERSION:-"develop"}
 
-PANTRY_CONTENT_URL="http://pantry.learningequality.org/downloads/ka-lite/$VERSION/content"
+echo "KA-Lite OS X build script for version '$VERSION' and above."
+
+CONTENT_VERSION="0.17"
+PANTRY_CONTENT_URL="http://pantry.learningequality.org/downloads/ka-lite/$CONTENT_VERSION/content"
 
 
 ((STEP++))
@@ -97,7 +103,7 @@ echo "$STEP/$STEPS. Checking the arguments..."
 # this will make it easier to "rename" the archive.
 # Example: KA_LITE_REPO_ZIP="https://github.com/learningequality/ka-lite/archive/develop.zip"
 # KA_LITE_REPO_ZIP="https://github.com/learningequality/ka-lite/archive/develop.zip"
-KA_LITE_REPO_ZIP="https://github.com/learningequality/ka-lite/archive/$VERSION.x.zip"
+KA_LITE_REPO_ZIP="https://github.com/learningequality/ka-lite/archive/$VERSION.zip"
 
 
 # Check if an argument was passed as URL for the script and use that instead.
@@ -146,7 +152,7 @@ fi
 CONTENTPACKS_DIR="$WORKING_DIR/content/contentpacks"
 test ! -d "$CONTENTPACKS_DIR" && mkdir -p "$CONTENTPACKS_DIR"
 
-CONTENTPACKS_EN_PATH="$CONTENTPACKS_DIR/$CONTENTPACKS_EN_ZIP"
+CONTENTPACKS_EN_PATH="$CONTENTPACKS_DIR/contentpack-$CONTENT_VERSION.en.zip"
 echo "$STEP/$STEPS. Checking for en.zip"
 if [ -f "$CONTENTPACKS_EN_PATH" ]; then
     echo ".. Found '$CONTENTPACKS_EN_PATH' so will not re-download.  Delete it to re-download."
@@ -159,6 +165,47 @@ else
     fi
 fi
 
+# Create output directory path.
+OUTPUT_PATH="$WORKING_DIR/output"
+TEMP_OUTPUT_PATH="$WORKING_DIR/temp-output"
+test -e "$TEMP_OUTPUT_PATH" && rm -r "$TEMP_OUTPUT_PATH"
+test -e "$OUTPUT_PATH" && rm -r "$OUTPUT_PATH"
+test ! -d "$OUTPUT_PATH" && mkdir "$OUTPUT_PATH"
+test ! -d "$TEMP_OUTPUT_PATH" && mkdir "$TEMP_OUTPUT_PATH"
+
+# Download python installer.
+((STEP++))
+PYTHON_DOWNLOAD_URL="https://www.python.org/ftp/python/2.7.12/python-2.7.12-macosx10.6.pkg"
+cd $TEMP_OUTPUT_PATH
+echo "$STEP/$STEPS. Downloading the minimum requirement of Python Installer at $PYTHON_DOWNLOAD_URL ..."
+wget --retry-connrefused --read-timeout=20 --waitretry=1 -t 100 --continue $PYTHON_DOWNLOAD_URL
+if [ $? -ne 0 ]; then
+    echo ".. Abort!  Can't download Python at '$PYTHON_DOWNLOAD_URL'"
+    exit 1
+fi
+
+KA_LITE_PROJECT_DIR="$SCRIPTPATH/KA-Lite"
+DMG_PATH="$OUTPUT_PATH/KA-Lite-Installer.dmg"
+DMG_BUILDER_PATH="$WORKING_DIR/create-dmg"
+CREATE_DMG="$DMG_BUILDER_PATH/create-dmg"
+KA_LITE_ICNS_PATH="$KA_LITE_PROJECT_DIR/KA-Lite/Resources/images/ka-lite.icns"
+
+test ! -d "$OUTPUT_PATH" && mkdir "$OUTPUT_PATH"
+
+((STEP++))
+echo "$STEP/$STEPS. Checking create dmg library..."
+CREATE_DMG_ZIP="$WORKING_DIR/create-dmg.zip"
+CREATE_DMG_URL="https://github.com/mrpau/create-dmg/archive/master.zip"
+# clone the .dmg builder if non-existent
+if ! [ -d $DMG_BUILDER_PATH ]; then
+    cd $WORKING_DIR
+    echo ".. Downloading create dmg library at '$CREATE_DMG_URL'..."
+    wget --retry-connrefused --read-timeout=20 --waitretry=1 -t 100 --continue -O $CREATE_DMG_ZIP $CREATE_DMG_URL
+        # Extract KA-Lite
+    echo ".. Extracting '$CREATE_DMG_ZIP'..."
+    tar -xf $CREATE_DMG_ZIP -C $WORKING_DIR
+    mv $WORKING_DIR/create-dmg-* create-dmg
+fi
 
 ((STEP++))
 echo "$STEP/$STEPS. Checking Github source..."
@@ -202,66 +249,47 @@ else
     fi
 fi
 
-
 ((STEP++))
-echo "$STEP/$STEPS. Checking Pyrun..."
+echo "$STEP/$STEPS. Install and create virtualenv..."
 
-# TODO(arceduardvincent): Update the pyrun url if the "Failed to install setuptools" issue is fix.
-INSTALL_PYRUN_URL="https://downloads.egenix.com/python/index/ucs2/egenix-pyrun/2.2.0/install-pyrun?filename=install-pyrun"
-INSTALL_PYRUN="$WORKING_DIR/install-pyrun.sh"
-PYRUN_NAME="pyrun-2.7"
-PYRUN_DIR="$WORKING_DIR/$PYRUN_NAME"
-PYRUN_BIN="$PYRUN_DIR/bin"
-PYRUN="$PYRUN_BIN/pyrun"
-PYRUN_PIP="$PYRUN_BIN/pip"
+# Must use Python version 2.7.11+ to build KA Lite. 
+PIP_CMD="pip install virtualenv"
+cd "$KA_LITE_DIR"
+echo ".. Running $PIP_CMD..."
+$PIP_CMD
 
-# Don't download Pyrun if there's already a `pyrun-2.7` directory.
-if [ -d "$PYRUN_DIR" ]; then
-    echo ".. Found PyRun directory at '$PYRUN_DIR' so will not re-download.  Delete this folder to re-download."
-else
-    # Download install-pyrun
-    if [ -e "$INSTALL_PYRUN" ]; then
-        echo ".. Found '$INSTALL_PYRUN' so will not re-download.  Delete this file to re-download."
-    else
-        echo ".. Downloading 'install-pyrun' script..."
-        wget --retry-connrefused --read-timeout=20 --waitretry=1 -t 100 --continue -O $INSTALL_PYRUN $INSTALL_PYRUN_URL
-        if [ $? -ne 0 ]; then
-          echo ".. Abort!  Can't download 'install-pyrun' script."
-          exit 1
-        fi
-        chmod +x $INSTALL_PYRUN
-    fi
-
-    # Download PyRun.
-    echo ".. Downloading PyRun with Python 2.7..."
-    $INSTALL_PYRUN --python=2.7 $PYRUN_DIR
-    if [ $? -ne 0 ]; then
-        echo ".. Abort!  Can't install minimal PyRun."
-        exit 1
-    fi
+bash $SCRIPTPATH/ka-lite-python-version-check.sh
+if [ $? -ne 0 ]; then
+    echo ".. Abort! Python 2.7.11+ is required to build KA Lite"
+    exit 1
 fi
 
-# MUST: Override the PATH to add the path to the Pyrun binaries first so it's python executes instead of
-# the system python.  When the script exits the old PATH values will be restored.
-export PATH="$PYRUN_BIN:$PATH"
+VENV_PATH="$(which virtualenv)"
+ENV_CMD="$VENV_PATH venv --python=python2.7"
+$ENV_CMD
+if [ $? -ne 0 ]; then
+    echo ".. Abort!  Error/s encountered running $ENV_CMD"
+    exit 1
+fi
 
+source venv/bin/activate
+ENV_PATH="$(pwd venv)"
 
 ((STEP++))
-echo "$STEP/$STEPS. Upgrading Pyrun's Pip..."
+echo "$STEP/$STEPS. Upgrading Python Pip..."
 
-# MUST: Upgrade Pyrun's pip from v1.5.6 to prevent issues.
-UPGRADE_PIP_CMD="$PYRUN_PIP install --upgrade pip"
+# MUST: Upgrade Python's pip.
+UPGRADE_PIP_CMD="pip install --upgrade pip"
 $UPGRADE_PIP_CMD
 if [ $? -ne 0 ]; then
     echo ".. Abort!  Error/s encountered running '$UPGRADE_PIP_CMD'."
     exit 1
 fi
 
-
 ((STEP++))
 echo "$STEP/$STEPS. Installing Pip requirements for use of Makefile..."
 
-PIP_CMD="$PYRUN_PIP install -r requirements_dev.txt"
+PIP_CMD="pip install -r requirements_dev.txt"
 # TODO(cpauya): Streamline this to pip install only the needed modules/executables for `make dist` below.
 cd "$KA_LITE_DIR"
 echo ".. Running $PIP_CMD..."
@@ -271,23 +299,48 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-PIP_CMD="$PYRUN_PIP install -r requirements_sphinx.txt"
+PIP_CMD="pip install -r requirements_sphinx.txt"
 # TODO(cpauya): Streamline this to pip install only the needed modules/executables for `make dist` below.
 cd "$KA_LITE_DIR"
 echo ".. Running $PIP_CMD..."
 $PIP_CMD
 if [ $? -ne 0 ]; then
-    echo ".. Abort!  Error/s encountered running '$PIP_CMD'."
+    echo ".. Abort! Error/s encountered running '$PIP_CMD'."
     exit 1
 fi
 
+NPM_CMD="npm install"
+cd "$KA_LITE_DIR"
+echo ".. Running $NPM_CMD..."
+$NPM_CMD
+if [ $? -ne 0 ]; then
+    echo ".. Abort! Error/s encountered running '$NPM_CMD'."
+    exit 1
+fi
+
+PIP_CMD="pip install ."
+cd "$KA_LITE_DIR"
+echo ".. Running $PIP_CMD..."
+$PIP_CMD
+if [ $? -ne 0 ]; then
+    echo ".. Abort! Error/s encountered running '$PIP_CMD'."
+    exit 1
+fi
+
+PIP_CMD="pip install wheel"
+echo ".. Running $PIP_CMD..."
+$PIP_CMD
+if [ $? -ne 0 ]; then
+    echo ".. Abort! Error/s encountered running '$PIP_CMD'."
+    exit 1
+fi
 
 ((STEP++))
 echo "$STEP/$STEPS. Running 'make dist'..."
 
-# MUST: Make sure we have a KALITE_PYTHON env var that points to Pyrun
+# MUST: Make sure we have a KALITE_PYTHON env var that points to python
 # because `bin/kalite manage ...` will be called when we do `make assets`.
-export KALITE_PYTHON="$PYRUN"
+export KALITE_PYTHON="python"
 
 cd "$KA_LITE_DIR"
 MAKE_CMD="make dist"
@@ -298,17 +351,41 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-
 ((STEP++))
-echo "$STEP/$STEPS. Running 'setup.py install --static'..."
+echo "$STEP/$STEPS. Installing PEX to create kalite PEX file"
+PIP_CMD="pip install pex"
+echo ".. Running $PIP_CMD..."
+$PIP_CMD
+if [ $? -ne 0 ]; then
+    echo ".. Abort! Error/s encountered running '$PIP_CMD'."
+    exit 1
+fi
 
 cd "$KA_LITE_DIR"
-SETUP_CMD="$PYRUN setup.py install"
-SETUP_STATIC_CMD="$SETUP_CMD --static"
-echo ".. Running $SETUP_STATIC_CMD..."
-$SETUP_STATIC_CMD
+WHL_FILE="$(find dist/ -name 'ka_lite_static-*.whl')"
+pex -o dist/kalite.pex -m kalite $WHL_FILE
 if [ $? -ne 0 ]; then
-    echo ".. Abort!  Error/s encountered running '$SETUP_STATIC_CMD'."
+    echo ".. Abort! Failed to build KA Lite pex file."
+    exit 1
+fi
+
+((STEP++))
+echo "Updating KA Lite installer version"
+KA_LITE_VERSION="$(kalite --version)"
+DOCS_PATH="$SCRIPTPATH/KA-Lite-Packages/docs"
+sed -i '' "13s/{{ KA_LITE_VERSION }}/$KA_LITE_VERSION/g" "$DOCS_PATH/INTRODUCTION.rst.rtfd/TXT.rtf" && \
+sed -i '' "12s/{{ KA_LITE_VERSION }}/$KA_LITE_VERSION/g" "$DOCS_PATH/SUMMARY.rtfd/TXT.rtf" && \
+sed -i '' "12s/{{ KA_LITE_VERSION }}/$KA_LITE_VERSION/g" "$DOCS_PATH/README.rst.rtf"
+if [ $? -ne 0 ]; then
+    echo ".. Failed to update KA Lite installer version."
+fi
+
+ENV_CMD="rm -r $ENV_PATH/venv"
+deactivate
+echo ".. Removing $ENV_CMD..."
+$ENV_CMD
+if [ $? -ne 0 ]; then
+    echo ".. Abort!  Error/s encountered running '$ENV_CMD'."
     exit 1
 fi
 
@@ -316,7 +393,6 @@ fi
 # Build the Xcode project
 ((STEP++))
 echo "$STEP/$STEPS. Building the Xcode project..."
-KA_LITE_PROJECT_DIR="$SCRIPTPATH/KA-Lite"
 if [ -d "$KA_LITE_PROJECT_DIR" ]; then
     # MUST: xcodebuild needs to be on the same directory as the .xcodeproj file
     cd "$KA_LITE_PROJECT_DIR"
@@ -326,6 +402,7 @@ if [ -d "$KA_LITE_PROJECT_DIR" ]; then
         exit 1
     fi
 fi
+
 # check if build of Xcode project succeeded
 KA_LITE_APP_PATH="$KA_LITE_PROJECT_DIR/build/Release/KA-Lite.app"
 if ! [ -d "$KA_LITE_APP_PATH" ]; then
@@ -334,14 +411,14 @@ if ! [ -d "$KA_LITE_APP_PATH" ]; then
 fi
 
 
-# Check if to codesign or not
+# Check if to code-sign or not
 ((STEP++))
-echo "$STEP/$STEPS. Checking if to codesign the application or not..."
+echo "$STEP/$STEPS. Checking if to code-sign the application or not..."
 SIGNER_IDENTITY_APPLICATION="Developer ID Application: Foundation for Learning Equality, Inc. (H83B64B6AV)"
 if [ -z ${IS_KALITE_RELEASE+0} ]; then 
-    echo ".. Not a release, don't codesign the application!"
+    echo ".. Not a release, don't code-sign the application!"
 else 
-    echo ".. Release build so MUST codesign the application..."
+    echo ".. Release build so MUST code-sign the application..."
     codesign -d -s "$SIGNER_IDENTITY_APPLICATION" --force "$KA_LITE_APP_PATH"
     if [ $? -ne 0 ]; then
         echo ".. Abort!  Error/s encountered codesigning '$KA_LITE_APP_PATH'."
@@ -353,9 +430,7 @@ fi
 # Build the KA-Lite  installer using `Packages` to generate the .pkg file.
 ((STEP++))
 cd "$WORKING_DIR/.."
-OUTPUT_PATH="$WORKING_DIR/output"
-echo "$STEP/$STEPS. Building the .pkg file at '$OUTPUT_PATH'..."
-test ! -d "$OUTPUT_PATH" && mkdir "$OUTPUT_PATH"
+echo "$STEP/$STEPS. Building the .pkg file at '$TEMP_OUTPUT_PATH'..."
 
 KALITE_PACKAGES_NAME="KA-Lite.pkg"
 PACKAGES_PROJECT="$SCRIPTPATH/KA-Lite-Packages/KA-Lite.pkgproj"
@@ -368,11 +443,11 @@ if [ $? -ne 0 ]; then
 fi
 
 echo ".. Checking if to productsign the package or not..."
-OUTPUT_PKG="$OUTPUT_PATH/$KALITE_PACKAGES_NAME"
+OUTPUT_PKG="$TEMP_OUTPUT_PATH/$KALITE_PACKAGES_NAME"
 SIGNER_IDENTITY_INSTALLER="Developer ID Installer: Foundation for Learning Equality, Inc. (H83B64B6AV)"
 if [ -z ${IS_KALITE_RELEASE+0} ]; then 
     echo ".. Not a release, don't productsign the package!"
-    mv -v $PACKAGES_OUTPUT $OUTPUT_PATH
+    mv -v $PACKAGES_OUTPUT $TEMP_OUTPUT_PATH
 else 
     echo ".. Release build so MUST productsign the package..."
     productsign --sign "$SIGNER_IDENTITY_INSTALLER" "$PACKAGES_OUTPUT" "$OUTPUT_PKG"
@@ -382,7 +457,47 @@ else
     fi
 fi
 
+# Remove the .dmg if it exists.
+test -e "$DMG_PATH" && rm "$DMG_PATH"
 
-echo "Congratulations! Your newly built installer is at '$OUTPUT_PKG'."
-cd "$WORKING_DIR/.."
+# Add the README.md to the package.
+# Copy the KA Lite logo to the package.
+cp "$SCRIPTPATH/dmg-resources/README.md" "$TEMP_OUTPUT_PATH"
+cp "$SCRIPTPATH/dmg-resources/ka-lite-logo.png" "$TEMP_OUTPUT_PATH"
+
+echo "$STEP/$STEPS. Building the .dmg file at '$OUTPUT_PATH'..."
+# Let's create the .dmg.
+$CREATE_DMG \
+    --volname "KA Lite Installer" \
+    --volicon "$KA_LITE_ICNS_PATH" \
+    --background "$TEMP_OUTPUT_PATH/ka-lite-logo.png" \
+    --icon "KA-Lite.pkg" 100 170 \
+    --icon "README.md" 300 170 \
+    --icon "python-2.7.12-macosx10.6.pkg" 500 170 \
+    --window-size 600 400 \
+    "$DMG_PATH" \
+    "$TEMP_OUTPUT_PATH"
+if [ $? -ne 0 ]; then
+    echo ".. Abort! Failed to build KA Lite dmg file."
+    exit 1
+fi
+
 echo "Done!"
+if [ -e "$DMG_PATH" ]; then
+    # codesign the built DMG file
+    # unlock the keychain first so we can access the private key
+    # security unlock-keychain -p $KEYCHAIN_PASSWORD
+    if [ -z ${IS_KALITE_RELEASE+0} ]; then 
+        echo "Congratulations! Your newly built KA Lite installer is at '$DMG_PATH'."
+    else
+        codesign -s "$SIGNER_IDENTITY_APPLICATION" --force "$DMG_PATH"
+        if [ $? -ne 0 ]; then
+            echo "..Failed to codesign the newly built KA Lite installer at '$DMG_PATH'."
+            exit 1
+        fi
+        echo "Congratulations! Your newly built KA Lite installer is at '$DMG_PATH'."
+    fi
+else
+    echo "Sorry, something went wrong trying to build the installer at '$DMG_PATH'."
+    exit 1
+fi